requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
frozenlist==1.4.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
magic-filter==1.0.12
matplotlib==3.8.2
multidict==6.0.4
openai==0.28.1
pydantic==2.3.0
python-telegram-bot==20.6
requests==2.31.0
seaborn==0.13.0
sniffio==1.3.0
translators==5.8.9
tqdm==4.66.1
urllib3==2.0.6
yarl==1.9.2
pytest==7.4.3
